organise data files more logically
