+++ app/javascript/pages/country/data/indicators.json
