+++ app/javascript/pages/country/data/thresholds.json
