+++ app/javascript/pages/country/widgets/widget-total-area-plantations/widget-total-area-plantations.js
@@ -1,7 +1,7 @@
 import { createElement } from 'react';
 import { connect } from 'react-redux';
 
-import { getUnits } from 'pages/country/utils/filters';
+import { getUnits } from 'pages/country/widgets/widgets-selectors';
 
 import WidgetTotalAreaPlantationsComponent from './widget-total-area-plantations-component';
 import actions from './widget-total-area-plantations-actions';
