+++ app/javascript/pages/country/widgets/widget-tree-loss/widget-tree-loss-component.js
@@ -57,7 +57,7 @@ class WidgetTreeLoss extends PureComponent {
             onEndYearChange={setTreeLossSettingsEndYear}
           />
         </WidgetHeader>
-        {isLoading ? (
+        {!isLoading ? (
           <Loader className="loader-offset" />
         ) : (
           <div>
